Separated 3PCF post utility function
In preparation for separate mock post-processing
